edits
Hey @emileighoutlaw - can you take a look at the edits above?  Because of the uniqueness of the JV authority, we want to make it somehow clear that it's best for projects that don't have good alternatives elsewhere. 

See #5 here as an example - https://www.ntis.gov/assets/pdf/FAQsFull_FINAL_070816.pdf